refactor: 관심사별 패키지화
