+++ docs/README.md
@@ -97,5 +97,5 @@
 - [x] 플레이어의 배팅 금액을 저장할 수 있다.
 - [x] 배팅 금액은 1000 ~ 10000원 사이여야 한다.
 - [x] 배팅 금액은 1000원 단위여야 한다.
-- [ ] 딜러 수익을 알 수 있다.
-- [ ] 플레이어들의 수익을 알 수 있다.
+- [x] 딜러 수익을 알 수 있다.
+- [x] 플레이어들의 수익을 알 수 있다.
