+++ src/main/java/blackjack/domain/card/CardRank.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 public enum CardRank {
