+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,5 +1,6 @@
-package blackjack.domain;
+package blackjack.domain.participant;
 
+import blackjack.domain.card.Card;
 import java.util.List;
 
 public abstract class Participant {
