+++ src/main/java/blackjack/view/OutputView.java
@@ -1,9 +1,9 @@
 package blackjack.view;
 
-import blackjack.domain.Card;
-import blackjack.domain.Dealer;
-import blackjack.domain.Participant;
-import blackjack.domain.Player;
+import blackjack.domain.card.Card;
+import blackjack.domain.participant.Dealer;
+import blackjack.domain.participant.Participant;
+import blackjack.domain.participant.Player;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
