+++ src/test/java/blackjack/domain/CardTest.java
@@ -2,6 +2,9 @@ package blackjack.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardRank;
+import blackjack.domain.card.CardShape;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
