+++ src/test/java/blackjack/domain/ParticipantsTest.java
@@ -3,6 +3,7 @@ package blackjack.domain;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import blackjack.domain.participant.Participants;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
